chore: rename test
